Fix building on Linux
Linux needed a bunch more libraries linked, some which are not available
on Windows, and they need to be included in a certain order, hence the
somewhat ugly solution in FindDucktape.cmake.

When building on Linux the libraries glfw3 and assimp-5 builds their
shared objects without the version number on the end. So I had to change
any linker references `glfw3 -> glfw` and `assimp-5 -> assimp`.